Develop

+++ server/src/main/java/com/orientechnologies/orient/server/network/protocol/binary/OBinaryNetworkProtocolAbstract.java
@@ -17,21 +17,19 @@ package com.orientechnologies.orient.server.network.protocol.binary;
 
 import java.io.IOException;
 import java.net.Socket;
-import java.net.SocketException;
-import java.util.List;
 import java.util.logging.Level;
 
-import com.orientechnologies.common.concur.lock.OLockException;
 import com.orientechnologies.common.exception.OException;
 import com.orientechnologies.common.log.OLogManager;
 import com.orientechnologies.orient.core.Orient;
 import com.orientechnologies.orient.core.config.OContextConfiguration;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.db.ODatabase.STATUS;
+import com.orientechnologies.orient.core.db.ODatabaseRecordThreadLocal;
+import com.orientechnologies.orient.core.db.document.ODatabaseDocument;
 import com.orientechnologies.orient.core.db.document.ODatabaseDocumentTx;
 import com.orientechnologies.orient.core.db.record.ODatabaseRecord;
 import com.orientechnologies.orient.core.db.record.OIdentifiable;
-import com.orientechnologies.orient.core.engine.local.OEngineLocal;
 import com.orientechnologies.orient.core.engine.local.OEngineLocalPaginated;
 import com.orientechnologies.orient.core.engine.memory.OEngineMemory;
 import com.orientechnologies.orient.core.exception.ODatabaseException;
@@ -41,11 +39,17 @@ import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.id.ORecordId;
 import com.orientechnologies.orient.core.metadata.security.ORole;
 import com.orientechnologies.orient.core.metadata.security.OUser;
+import com.orientechnologies.orient.core.record.ORecord;
 import com.orientechnologies.orient.core.record.ORecordInternal;
 import com.orientechnologies.orient.core.record.impl.ODocument;
+import com.orientechnologies.orient.core.serialization.serializer.ONetworkThreadLocalSerializer;
+import com.orientechnologies.orient.core.serialization.serializer.record.ORecordSerializer;
+import com.orientechnologies.orient.core.serialization.serializer.record.ORecordSerializerFactory;
 import com.orientechnologies.orient.core.serialization.serializer.record.OSerializationThreadLocal;
+import com.orientechnologies.orient.core.serialization.serializer.record.string.ORecordSerializerSchemaAware2CSV;
 import com.orientechnologies.orient.core.storage.OStorage;
-import com.orientechnologies.orient.core.storage.impl.local.OStorageLocalAbstract;
+import com.orientechnologies.orient.core.storage.OStorageProxy;
+import com.orientechnologies.orient.core.storage.impl.local.OAbstractPaginatedStorage;
 import com.orientechnologies.orient.core.version.ORecordVersion;
 import com.orientechnologies.orient.core.version.OVersionFactory;
 import com.orientechnologies.orient.enterprise.channel.OChannel;
@@ -53,6 +57,7 @@ import com.orientechnologies.orient.enterprise.channel.binary.OChannelBinaryProt
 import com.orientechnologies.orient.enterprise.channel.binary.OChannelBinaryServer;
 import com.orientechnologies.orient.enterprise.channel.binary.ONetworkProtocolException;
 import com.orientechnologies.orient.server.OServer;
+import com.orientechnologies.orient.server.network.OServerNetworkListener;
 import com.orientechnologies.orient.server.network.protocol.ONetworkProtocol;
 
 /**
@@ -62,16 +67,65 @@ import com.orientechnologies.orient.server.network.protocol.ONetworkProtocol;
  * 
  */
 public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
+  protected final Level          logClientExceptions;
+  protected final boolean        logClientFullStackTrace;
   protected OChannelBinaryServer channel;
   protected int                  requestType;
   protected int                  clientTxId;
-  protected final Level          logClientExceptions;
-  protected final boolean        logClientFullStackTrace;
 
   public OBinaryNetworkProtocolAbstract(final String iThreadName) {
     super(Orient.instance().getThreadGroup(), iThreadName);
     logClientExceptions = Level.parse(OGlobalConfiguration.SERVER_LOG_DUMP_CLIENT_EXCEPTION_LEVEL.getValueAsString());
     logClientFullStackTrace = OGlobalConfiguration.SERVER_LOG_DUMP_CLIENT_EXCEPTION_FULLSTACKTRACE.getValueAsBoolean();
+  }
+
+  @Override
+  public void config(final OServerNetworkListener iListener, final OServer iServer, final Socket iSocket,
+      final OContextConfiguration iConfig) throws IOException {
+    server = iServer;
+    channel = new OChannelBinaryServer(iSocket, iConfig);
+  }
+
+  @Override
+  public int getVersion() {
+    return OChannelBinaryProtocol.CURRENT_PROTOCOL_VERSION;
+  }
+
+  @Override
+  public void shutdown() {
+    channel.close();
+  }
+
+  @Override
+  public OChannel getChannel() {
+    return channel;
+  }
+
+  /**
+   * Write a OIdentifiable instance using this format:<br/>
+   * - 2 bytes: class id [-2=no record, -3=rid, -1=no class id, > -1 = valid] <br/>
+   * - 1 byte: record type [d,b,f] <br/>
+   * - 2 bytes: cluster id <br/>
+   * - 8 bytes: position in cluster <br/>
+   * - 4 bytes: record version <br/>
+   * - x bytes: record content <br/>
+   * 
+   * @param o
+   * @throws IOException
+   */
+  public void writeIdentifiable(final OIdentifiable o) throws IOException {
+    if (o == null)
+      channel.writeShort(OChannelBinaryProtocol.RECORD_NULL);
+    else if (o instanceof ORecordId) {
+      channel.writeShort(OChannelBinaryProtocol.RECORD_RID);
+      channel.writeRID((ORID) o);
+    } else {
+      writeRecord((ORecordInternal<?>) o.getRecord());
+    }
+  }
+
+  public String getType() {
+    return "binary";
   }
 
   /**
@@ -81,6 +135,8 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
    * @throws IOException
    */
   protected abstract boolean executeRequest() throws IOException;
+
+  protected abstract void sendError(final int iClientTxId, final Throwable t) throws IOException;
 
   /**
    * Executed before the request.
@@ -96,13 +152,6 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
    * @throws IOException
    */
   protected void onAfterRequest() throws IOException {
-  }
-
-  @Override
-  public void config(final OServer iServer, final Socket iSocket, final OContextConfiguration iConfig,
-      final List<?> iStatelessCommands, List<?> iStatefulCommands) throws IOException {
-    server = iServer;
-    channel = new OChannelBinaryServer(iSocket, iConfig);
   }
 
   @Override
@@ -148,117 +197,14 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
     }
   }
 
-  @Override
-  public void shutdown() {
-    channel.close();
-  }
-
-  @Override
-  public OChannel getChannel() {
-    return channel;
-  }
-
   protected void sendOk(final int iClientTxId) throws IOException {
     channel.writeByte(OChannelBinaryProtocol.RESPONSE_STATUS_OK);
     channel.writeInt(iClientTxId);
   }
 
-  protected void sendError(final int iClientTxId, final Throwable t) throws IOException {
-    channel.acquireWriteLock();
-    try {
-
-      channel.writeByte(OChannelBinaryProtocol.RESPONSE_STATUS_ERROR);
-      channel.writeInt(iClientTxId);
-
-      Throwable current;
-      if (t instanceof OLockException && t.getCause() instanceof ODatabaseException)
-        // BYPASS THE DB POOL EXCEPTION TO PROPAGATE THE RIGHT SECURITY ONE
-        current = t.getCause();
-      else
-        current = t;
-
-      while (current != null) {
-        // MORE DETAILS ARE COMING AS EXCEPTION
-        channel.writeByte((byte) 1);
-
-        channel.writeString(current.getClass().getName());
-        channel.writeString(current != null ? current.getMessage() : null);
-
-        current = current.getCause();
-      }
-      channel.writeByte((byte) 0);
-
-      channel.flush();
-
-      if (OLogManager.instance().isLevelEnabled(logClientExceptions)) {
-        if (logClientFullStackTrace)
-          OLogManager.instance().log(this, logClientExceptions, "Sent run-time exception to the client %s: %s", t,
-              channel.socket.getRemoteSocketAddress(), t.toString());
-        else
-          OLogManager.instance().log(this, logClientExceptions, "Sent run-time exception to the client %s: %s", null,
-              channel.socket.getRemoteSocketAddress(), t.toString());
-      }
-    } catch (Exception e) {
-      if (e instanceof SocketException)
-        shutdown();
-    } finally {
-      channel.releaseWriteLock();
-    }
-  }
-
-  /**
-   * Write a OIdentifiable instance using this format:<br/>
-   * - 2 bytes: class id [-2=no record, -3=rid, -1=no class id, > -1 = valid] <br/>
-   * - 1 byte: record type [d,b,f] <br/>
-   * - 2 bytes: cluster id <br/>
-   * - 8 bytes: position in cluster <br/>
-   * - 4 bytes: record version <br/>
-   * - x bytes: record content <br/>
-   * 
-   * @param o
-   * @throws IOException
-   */
-  public void writeIdentifiable(final OIdentifiable o) throws IOException {
-    if (o == null)
-      channel.writeShort(OChannelBinaryProtocol.RECORD_NULL);
-    else if (o instanceof ORecordId) {
-      channel.writeShort(OChannelBinaryProtocol.RECORD_RID);
-      channel.writeRID((ORID) o);
-    } else {
-      writeRecord((ORecordInternal<?>) o.getRecord());
-    }
-  }
-
-  private void writeRecord(final ORecordInternal<?> iRecord) throws IOException {
-    channel.writeShort((short) 0);
-    channel.writeByte(iRecord.getRecordType());
-    channel.writeRID(iRecord.getIdentity());
-    channel.writeVersion(iRecord.getRecordVersion());
-    try {
-      final byte[] stream = iRecord.toStream();
-
-      // TRIM TAILING SPACES (DUE TO OVERSIZE)
-      int realLength = stream.length;
-      for (int i = stream.length - 1; i > -1; --i) {
-        if (stream[i] == 32)
-          --realLength;
-        else
-          break;
-      }
-
-      channel.writeBytes(stream, realLength);
-    } catch (Exception e) {
-      channel.writeBytes(null);
-      final String message = "Error on unmarshalling record " + iRecord.getIdentity().toString() + " (" + e + ")";
-      OLogManager.instance().error(this, message, e);
-
-      throw new OSerializationException(message, e);
-    }
-  }
-
   protected void checkStorageExistence(final String iDatabaseName) {
     for (OStorage stg : Orient.instance().getStorages()) {
-      if (stg.getName().equalsIgnoreCase(iDatabaseName) && stg.exists())
+      if (!(stg instanceof OStorageProxy) && stg.getName().equalsIgnoreCase(iDatabaseName) && stg.exists())
         throw new ODatabaseException("Database named '" + iDatabaseName + "' already exists: " + stg);
     }
   }
@@ -280,7 +226,7 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
     }
 
     OLogManager.instance().info(this, "Created database '%s' of type '%s'", iDatabase.getName(),
-        iDatabase.getStorage() instanceof OStorageLocalAbstract ? iDatabase.getStorage().getType() : "memory");
+        iDatabase.getStorage() instanceof OAbstractPaginatedStorage ? iDatabase.getStorage().getType() : "memory");
 
     // if (iDatabase.getStorage() instanceof OStorageLocal)
     // // CLOSE IT BECAUSE IT WILL BE OPEN AT FIRST USE
@@ -295,7 +241,7 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
     final OStorage stg = Orient.instance().getStorage(dbName);
     if (stg != null)
       path = stg.getURL();
-    else if (storageType.equals(OEngineLocal.NAME) || storageType.equals(OEngineLocalPaginated.NAME)) {
+    else if (storageType.equals(OEngineLocalPaginated.NAME)) {
       // if this storage was configured return always path from config file, otherwise return default path
       path = server.getConfiguration().getStoragePath(dbName);
 
@@ -311,9 +257,23 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
 
   protected int deleteRecord(final ODatabaseRecord iDatabase, final ORID rid, final ORecordVersion version) {
     try {
+      // TRY TO SEE IF THE RECORD EXISTS
+      final ORecord<?> record = rid.getRecord();
+      if (record == null)
+        return 0;
+
       iDatabase.delete(rid, version);
       return 1;
     } catch (Exception e) {
+      return 0;
+    }
+  }
+
+  protected int hideRecord(final ODatabaseRecord iDatabase, final ORID rid) {
+    try {
+      iDatabase.hide(rid);
+      return 1;
+    } catch (ORecordNotFoundException e) {
       return 0;
     }
   }
@@ -324,23 +284,19 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
   }
 
   protected ORecordInternal<?> createRecord(final ODatabaseRecord iDatabase, final ORecordId rid, final byte[] buffer,
-      final byte recordType, final int dataSegmentId) {
+      final byte recordType) {
     final ORecordInternal<?> record = Orient.instance().getRecordFactoryManager().newInstance(recordType);
-    record.fill(rid, OVersionFactory.instance().createVersion(), buffer, true);
-    if (dataSegmentId > 0)
-      record.setDataSegmentName(iDatabase.getDataSegmentNameById(dataSegmentId));
+    fillRecord(rid, buffer, OVersionFactory.instance().createVersion(), record, iDatabase);
     iDatabase.save(record);
     return record;
   }
 
   protected ORecordVersion updateRecord(final ODatabaseRecord iDatabase, final ORecordId rid, final byte[] buffer,
-      final ORecordVersion version, final byte recordType) {
+      final ORecordVersion version, final byte recordType, boolean updateContent) {
     final ORecordInternal<?> newRecord = Orient.instance().getRecordFactoryManager().newInstance(recordType);
-    newRecord.fill(rid, version, buffer, true);
+    fillRecord(rid, buffer, version, newRecord, null);
 
-    // if (((OSchemaProxy) iDatabase.getMetadata().getSchema()).getIdentity().equals(rid))
-    // // || ((OIndexManagerImpl) connection.database.getMetadata().getIndexManager()).getDocument().getIdentity().equals(rid)) {
-    // throw new OSecurityAccessException("Cannot update internal record " + rid);
+    newRecord.setContentChanged(updateContent);
 
     final ORecordInternal<?> currentRecord;
     if (newRecord instanceof ODocument) {
@@ -372,4 +328,84 @@ public abstract class OBinaryNetworkProtocolAbstract extends ONetworkProtocol {
     } catch (IOException e1) {
     }
   }
+
+  private void writeRecord(final ORecordInternal<?> iRecord) throws IOException {
+    channel.writeShort((short) 0);
+    channel.writeByte(iRecord.getRecordType());
+    channel.writeRID(iRecord.getIdentity());
+    channel.writeVersion(iRecord.getRecordVersion());
+    try {
+      final byte[] stream = getRecordBytes(iRecord);
+
+      int realLength = stream.length;
+      // TODO: This Logic should not be here provide an api in the Serializer for ask for trimmed content.
+      final ODatabaseRecord db = ODatabaseRecordThreadLocal.INSTANCE.getIfDefined();
+      if (db != null && db instanceof ODatabaseDocument) {
+        if (db.getSerializer() instanceof ORecordSerializerSchemaAware2CSV) {
+          // TRIM TAILING SPACES (DUE TO OVERSIZE)
+          for (int i = stream.length - 1; i > -1; --i) {
+            if (stream[i] == 32)
+              --realLength;
+            else
+              break;
+          }
+
+        }
+      }
+
+      channel.writeBytes(stream, realLength);
+    } catch (Exception e) {
+      channel.writeBytes(null);
+      final String message = "Error on unmarshalling record " + iRecord.getIdentity().toString() + " (" + e + ")";
+      OLogManager.instance().error(this, message, e);
+
+      throw new OSerializationException(message, e);
+    }
+  }
+
+  public void fillRecord(final ORecordId rid, final byte[] buffer, final ORecordVersion version, final ORecordInternal<?> record,
+      ODatabaseRecord iDatabase) {
+    String dbSerializerName = "";
+    if (iDatabase != null)
+      dbSerializerName = iDatabase.getSerializer().toString();
+
+    String name = getRecordSerializerName();
+    if (record.getRecordType() == ODocument.RECORD_TYPE && !dbSerializerName.equals(name)) {
+      record.fill(rid, version, null, true);
+      try {
+        ORecordSerializer ser = ORecordSerializerFactory.instance().getFormat(name);
+        ONetworkThreadLocalSerializer.setNetworkSerializer(ser);
+        record.fromStream(buffer);
+      } finally {
+        ONetworkThreadLocalSerializer.setNetworkSerializer(null);
+      }
+      record.setDirty();
+    } else
+      record.fill(rid, version, buffer, true);
+  }
+
+  protected byte[] getRecordBytes(final ORecordInternal<?> iRecord) {
+
+    final byte[] stream;
+    try {
+      String dbSerializerName = null;
+      if (ODatabaseRecordThreadLocal.INSTANCE.getIfDefined() != null)
+        dbSerializerName = iRecord.getDatabase().getSerializer().toString();
+      String name = getRecordSerializerName();
+      if (iRecord.getRecordType() == ODocument.RECORD_TYPE && (dbSerializerName == null || !dbSerializerName.equals(name))) {
+        ORecordSerializer ser = ORecordSerializerFactory.instance().getFormat(dbSerializerName);
+        ONetworkThreadLocalSerializer.setNetworkSerializer(ser);
+        ((ODocument) iRecord).deserializeFields();
+        ser = ORecordSerializerFactory.instance().getFormat(name);
+        ONetworkThreadLocalSerializer.setNetworkSerializer(ser);
+      }
+      stream = iRecord.toStream();
+    } finally {
+      ONetworkThreadLocalSerializer.setNetworkSerializer(null);
+    }
+    return stream;
+  }
+
+  protected abstract String getRecordSerializerName();
+
 }
